Split out probing and the building of probe targets.
