+++ pkg/reconciler/ingress/ingress_test.go
@@ -404,7 +404,7 @@ func TestReconcile(t *testing.T) {
 				config: ReconcilerTestConfig(),
 			},
 			statusManager: &fakeStatusManager{
-				FakeIsReady: func(ia *v1alpha1.Ingress, gw map[v1alpha1.IngressVisibility]sets.String) (bool, error) {
+				FakeIsReady: func(ctx context.Context, ia *v1alpha1.Ingress) (bool, error) {
 					return true, nil
 				},
 			},
@@ -818,7 +818,7 @@ func TestReconcile_EnableAutoTLS(t *testing.T) {
 				},
 			},
 			statusManager: &fakeStatusManager{
-				FakeIsReady: func(ia *v1alpha1.Ingress, gw map[v1alpha1.IngressVisibility]sets.String) (bool, error) {
+				FakeIsReady: func(ctx context.Context, ia *v1alpha1.Ingress) (bool, error) {
 					return true, nil
 				},
 			},
@@ -1007,7 +1007,7 @@ func newTestSetup(t *testing.T, configs ...*corev1.ConfigMap) (
 	controller := NewController(ctx, configMapWatcher)
 
 	controller.Reconciler.(*Reconciler).statusManager = &fakeStatusManager{
-		FakeIsReady: func(ia *v1alpha1.Ingress, gw map[v1alpha1.IngressVisibility]sets.String) (bool, error) {
+		FakeIsReady: func(ctx context.Context, ia *v1alpha1.Ingress) (bool, error) {
 			return true, nil
 		},
 	}
@@ -1239,9 +1239,9 @@ func makeGatewayMap(publicGateways []string, privateGateways []string) map[v1alp
 }
 
 type fakeStatusManager struct {
-	FakeIsReady func(ia *v1alpha1.Ingress, gw map[v1alpha1.IngressVisibility]sets.String) (bool, error)
+	FakeIsReady func(ctx context.Context, ia *v1alpha1.Ingress) (bool, error)
 }
 
-func (m *fakeStatusManager) IsReady(ia *v1alpha1.Ingress, gw map[v1alpha1.IngressVisibility]sets.String) (bool, error) {
-	return m.FakeIsReady(ia, gw)
+func (m *fakeStatusManager) IsReady(ctx context.Context, ia *v1alpha1.Ingress) (bool, error) {
+	return m.FakeIsReady(ctx, ia)
 }
